requirements.txt
Django==5.1.2
djangorestframework==3.15.2
requests==2.32.3
beautifulsoup4==4.12.3